+++ CHANGES.md
@@ -4,6 +4,8 @@
 
 ### Improvements
 
+-   Support writing dataframes without geometry column (#267)
+
 -   Calculate feature count by iterating over features if GDAL returns an
     unknown count for a data layer (e.g., OSM driver); this may have signficant
     performance impacts for some data sources that would otherwise return an